+++ app/src/main/java/com/hacybeyker/snapdoc/feature/camera/ui/CameraPermissionIntent.kt
@@ -1,0 +1,8 @@
+package com.hacybeyker.snapdoc.feature.camera.ui
+
+sealed interface CameraPermissionIntent {
+    data object RequestPermission : CameraPermissionIntent
+    data class PermissionResultReceived(val isGranted: Boolean, val shouldShowRationale: Boolean) :
+        CameraPermissionIntent
+    data object OpenAppSettings : CameraPermissionIntent
+}